doc(p2p): document the swarm and p2p implementations
