+++ crates/p2p/src/events.rs
@@ -1,13 +1,18 @@
+//! Events emitted from P2P.
+
 use prost::Message;
 use strata_p2p_wire::p2p::v1::GossipsubMsg;
 
 /// Events emitted from P2P to handle from operator side.
+///
+/// # Implementation Details
 ///
 /// `DepositSetupPayload` is generic as some implementation details for different BitVM2
 /// applications may vary. The only requirement for them to be decodable from bytes as protobuf
 /// message.
 #[derive(Clone, Debug)]
 pub enum Event<DepositSetupPayload: Message + Clone> {
+    /// Received message from other operator.
     ReceivedMessage(GossipsubMsg<DepositSetupPayload>),
 }
