+++ crates/p2p/src/lib.rs
@@ -1,3 +1,5 @@
+//! Strata P2P implementation.
+
 pub mod commands;
 pub mod events;
 pub mod swarm;
