+++ crates/p2p/src/swarm/handle.rs
@@ -9,13 +9,16 @@ use tokio::sync::{
 use crate::{commands::Command, events::Event};
 
 /// Handle to interact with P2P implementation spawned in another async
-/// task. To create new one, use [`super::P2P::new_handle`].
+/// task. To create a new one, use [`super::P2P::new_handle`].
 #[derive(Debug)]
 pub struct P2PHandle<DSP>
 where
     DSP: prost::Message + Clone,
 {
+    /// Event channel for the swarm.
     events: broadcast::Receiver<Event<DSP>>,
+
+    /// Command channel for the swarm.
     commands: mpsc::Sender<Command<DSP>>,
 }
 
@@ -23,6 +26,7 @@ impl<DSP> P2PHandle<DSP>
 where
     DSP: prost::Message + Clone,
 {
+    /// Creates a new [`P2PHandle`].
     pub(crate) fn new(
         events: broadcast::Receiver<Event<DSP>>,
         commands: mpsc::Sender<Command<DSP>>,
@@ -30,17 +34,17 @@ where
         Self { events, commands }
     }
 
-    /// Send command to P2P.
+    /// Sends command to P2P.
     pub async fn send_command(&self, command: impl Into<Command<DSP>>) {
         let _ = self.commands.send(command.into()).await;
     }
 
-    /// Get next event from P2P from events channel.
+    /// Gets the next event from P2P from events channel.
     pub async fn next_event(&mut self) -> Result<Event<DSP>, RecvError> {
         self.events.recv().await
     }
 
-    /// Check event's channel is empty or not.
+    /// Checks if the event's channel is empty or not.
     pub fn events_is_empty(&self) -> bool {
         self.events.is_empty()
     }
